+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 2.0.0
 
 - update to 1.21
-- eating a food item applies a configurable item cooldown to that item
+- eating a food item now applies a configurable cooldown to that item
+- fixed config option "food_effect_duration_threshold_to_allow_eating" not having an effect
 
 # 1.0.0